Create app_defualt
This is the default code that can be used and copied if any changes made brakes any default functions that should always work. 

This current version is still having issues with the FOLLOW and SPLIT upstream tactic and will update this when I have the whole file running as it should.