Update live athlete display
Updated code block that is responsible for the athlete profile tab. 

Updated the logic so does not loose the athletes you have selected when new data refreshes. 

Updated the logic so the the 1st upstream gate is not set to specific gates as 2 asnd 8, but rather is dynamic and adaptable, will now search and scan for the 1st upstream gate and and apply.